Remove redundant args and narrow scope
I moved the `--test` flag into the `register` subcommand scope as it doesn't seem to apply anywhere else.

I also removed the duplicate `--pkgs` arg from the `register` subcommand, as it is already defined at the main command level. Having them in two places makes it confusing as when each one applies will depend on the order of the args, e.g.

```
pyflyte --pkgs common.workflows serialize
```

will work, but

```
pyflyte --pkgs common.workflows register workflows
```

will not because the redefinition of `--pkgs` at the `register` subcommand level will override it and reset it back to the default. (You would have to put `--pkgs` in between register and workflows for it to apply for the register subcommand as it stands before this PR.)

This is unfortunately a breaking change though, as it means `--pkgs` is no longer allowed after `register`, and must be specified before.