Prefer the URL used to create the client object instead of configuration (#138)
When the RawSynchronousFlyteClient is created, it's given a host. We should prefer to use this instead of retrieving from configuration again.  Save the url passed into the client instead and when it comes time to get the Admin endpoints, use it first. Configuration fallback has been left in there in case we make the client url optional down the road.