pr into #1384 (#1407)
Signed-off-by: Yee Hing Tong <wild-endeavor@users.noreply.github.com>
Adds a test for false SSL for the old .ini style of config
Adds a shortcut for the pyflyte run/register commands - if no files are detected at all, create for sandbox instead. I think this is okay. The other option would be to fall back to checking the ~/.flyte/config-sandbox.yml file but I think using the code option is a bit cleaner, even though it ties us closer to the flytectl demo port settings.