Carry over hash from literal maps and collections in cache key computation
Until now, the hashes of literal maps and collections are not carried
over into the representative literal map/collection (respectively)
entering the cache key computation, despite the comments around the code
clearly stating so.

This commit enables carrying hashes from `LiteralMap`s and
`LiteralCollection`s over by explicitly setting the hashes on the
representer literal map/collection, respectively. It changes the loop
structure so that a substitute empty literal with the same hash is
returned in case a custom hash is set on a literal.

Also fixes a shadowing bug in the loops over the literal map/collection,
which happened since the bound loop argument and the input argument were
both named `literal`.

Adds a unit test verifying that the hashes are carried along for both
cases.

Signed-off-by: Nicholas Junge <nicholas.junge@web.de>