[DOM-68737] Refactor version_hash_additional_context to use bytes (#7)
## Why are the changes needed?

`bytes` is more flexible. This flexibility is relevant when including the serialized input bindings (bytes) in the version hash calculations. This is necessary to fix a bug that can occur:

There is a bug where Flyte rejects a valid workflow. This can happen when workflow inputs are read from a file, causing the workflow input bindings to be updated, without an update to the version.

## What changes were proposed in this pull request?

Refactor version_hash_additional_context, which currently uses `str`, to use `bytes`